+++ test-setup/jestFrameworkSetup.js
@@ -4,3 +4,11 @@
   @see https://facebook.github.io/jest/docs/configuration.html#setuptestframeworkscriptfile-string
 */
 require('jest-styled-components');
+
+if (!require.ensure) {
+  Object.defineProperties(Object.getPrototypeOf(require), {
+    ensure: {
+      value: (deps, cb) => cb(require),
+    },
+  });
+}